--- src/Lib.hs
@@ -1,4 +1,0 @@
-module Lib (someFunc) where
-
-someFunc :: IO ()
-someFunc = putStrLn "someFunc"
